components: fix accordion
