+++ components/accordion/sub-components/accordionItem.js
@@ -1,46 +1,43 @@
-import React, {useEffect, useRef} from "react";
+import React, { useEffect, useRef } from "react";
 import Heading from "../../heading";
 import StyledAccordionItem from "./styled-accordionItem";
 
 const AccordionItem = ({
-    heading,
-    level,
-    children,
-    isCollapsed,
-    onClick,
-    ...rest
-   
-  }) => {
-    const content = useRef();
+  heading,
+  level,
+  children,
+  isCollapsed,
+  onClick,
+  ...rest
 
-    return (
-      <StyledAccordionItem>
-        <div className="accordion">
-          <div className={`${isCollapsed ? "accordion__icon" : "accordion__icon rotate" }`} >+</div>
-          <Heading
-            onClick={onClick}
-            level={level}
-            style={{cursor: "pointer"}}
-          >{heading}
-          </Heading>
-        </div>
-        <div 
-          ref={content}
-          style={{ maxHeight: `${!isCollapsed ? `${content.current.scrollHeight}px` : "0px"}` }}
-          className="accordion__content"
-        >
-          <div className="accordion__text">{children}</div>
-        </div>
-      </StyledAccordionItem>
-    );
-  };
+}) => {
 
-  AccordionItem.propTypes = {
+  const content = useRef();
 
-  };
-  
-  AccordionItem.defaultProps = {
+  return (
+    <StyledAccordionItem {...rest}>
+      <div className="accordion">
+        <div className={`${isCollapsed ? "accordion__icon" : "accordion__icon rotate"}`} >+</div>
+        <Heading
+          onClick={onClick}
+          level={level}
+          style={{ cursor: "pointer" }}
+        >{heading}
+        </Heading>
+      </div>
+      <div
+        ref={content}
+        style={{ maxHeight: `${!isCollapsed ? `${content.current.scrollHeight}px` : "0px"}` }}
+        className="accordion__content"
+      >
+        <div className="accordion__text">{children}</div>
+      </div>
+    </StyledAccordionItem>
+  );
+};
 
-  };
-  
-  export default AccordionItem;
+AccordionItem.propTypes = {};
+
+AccordionItem.defaultProps = {};
+
+export default AccordionItem;
