+++ src/pages/index.js
@@ -4,7 +4,7 @@ import { useTranslation } from "gatsby-plugin-react-i18next";
 
 import Layout from "../../components/layout";
 import HeadSEO from "../sub-components/head-content";
-
+import { Accordion, AccordionItem } from "../../components/accordion";
 const IndexPage = () => {
 
   const {
@@ -18,7 +18,19 @@ const IndexPage = () => {
         <HeadSEO />
       </Layout.PageHead>
       <Layout.PageHeader></Layout.PageHeader>
-      <Layout.SectionMain></Layout.SectionMain>
+      <Layout.SectionMain>
+        <Accordion level={4}>
+          <AccordionItem heading="Are ONLYOFFICE forms free to use?">
+            Lorem ipsum
+          </AccordionItem>
+          <AccordionItem heading="Do I need to register to fill out a form?">
+            Lorem ipsum
+          </AccordionItem>
+          <AccordionItem heading="Can I download the forms and create documents using them locally?">
+            Lorem ipsum
+          </AccordionItem>
+        </Accordion>
+      </Layout.SectionMain>
       <Layout.PageFooter></Layout.PageFooter>
     </Layout>
   );
